Move tests unrelated to calculation elsewhere
