+++ tests/test_special_calculations.py
@@ -2,7 +2,6 @@ from decimal import Decimal
 
 import pytest
 from sympy import oo  # infinity in SymPy
-from sympy.core.symbol import Symbol
 from sympy.core.numbers import Rational
 
 from lenscalc import Lens
@@ -49,37 +48,6 @@ def test_different_refractive_index():
 
     assert compare_two_lenses(comparison_lens, lens)
     assert lens.NPS != 0
-
-
-def test_extra_symbol():
-    """
-    Test lens with an extra attribute of Symbol type.
-    """
-    lens = Lens()
-
-    extra = Symbol("extra")
-
-    # Set the arttribute.
-    lens.extra = extra
-    # Try to get the value of the attribute.
-    extra_attribute = lens.extra
-
-    assert extra_attribute == extra
-
-
-def test_extra_string():
-    """
-    Test lens with extra attributes of string type.
-
-    The attributes should not be converted to type float.
-    """
-    lens = Lens()
-
-    lens.extra_string_text = "extra"
-    lens.extra_string_number = "1.5"
-
-    assert isinstance(lens.extra_string_text, str)
-    assert isinstance(lens.extra_string_number, str)
 
 
 def test_planar_surface_lens():
